requirements.txt
aiogram
python-decouple
Pillow